pyproject.toml
[tool.poetry]
name = "repl_nix_GDSCWasedaSlackbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
slack-bolt = "^1.18.0"
python-dotenv = "^1.0.0"
google-auth = "^2.20.0"
google-auth-oauthlib = "^1.0.0"
cryptography = "^41.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
